feature(mark): AnalysisServer includes MarkModel.
AnalysisServer now includes a fat-Jar with a standalone MARK parser.
This jar is built by the mark-crymlin-eclipse-plugin project in parallel
to the existing Eclipse plugin. It is loaded from mavenLocal() and
marked as a "dynamic" dependency which requires gradle to re-load it
even if it already exists in gradle cache. This build will fail on the
build server until "de.fhg.aisec.mark:de.fhg.aisec.mark:1.0.0-SNAPSHOT"
is pushed to the netsec registry.