Use a custom formatter. The google formatter religiously enforces a line length 80, even for comments. This is not only a waste of space, but it messes up the code, e.g. in case of long method signatures with generics.
We now have a custom formatter XML file that is used by spotless and can be imported directly into IntelliJ and Eclipse. It allows a line length of 170. If you are using the IntelliJ "google formatter" plugin, make sure to disable it for this project and use the standard IntelliJ formatter instead. The settings may be imported from "formatter-settings.xml".