Adds a case resolving literals of referenced variables in expression lists
An expression list (C++) can contain variable references. These variables
are linked via DFG edges. If an intermittent assignment to a variable uses
a literal, the immediate DFG edge can point to a literal instead of a
variable declaration.

This adds a case that checks for a literal and processes its value.