Persist graphs to database using iterative procedure (#91)
* Remove increased stack size for JVM

We remove the JVM flag to increase the stack size, because we want
to see the problem with stack exhaustion due to excessive recursion.

* Simplify Database interface by removing unused return parameter

Database connect() returned a Boolean that was simply true at the end of the
method when no exception occured. This return value wasn't used anywhere.
Removed it to simplify interface.

* Degenerify OverflowDatabase

OverflowDatabase passed the generic type along. However, internally some method where expecting
that the generic type was of type Node. To simplify the code, the generics was removed and the
generic Database interface was initialized with the Node type.

In addition, javadoc comments where added.

* Removed generics of OverflowDatabase in test classes

The generic type of OverflowDatabase was removed due to type dependencies. This
updates the test classes to not use generics.

* Use generic interface instead of implementing class

Changed type of field to the generic interface instead of the implementing
class. This hides accessible methods in OverflowDatabase.

* Remove increased stack size for JVM

We remove the JVM flag to increase the stack size, because we want
to see the problem with stack exhaustion due to excessive recursion.

* Simplify Database interface by removing unused return parameter

Database connect() returned a Boolean that was simply true at the end of the
method when no exception occured. This return value wasn't used anywhere.
Removed it to simplify interface.

* Degenerify OverflowDatabase

OverflowDatabase passed the generic type along. However, internally some method where expecting
that the generic type was of type Node. To simplify the code, the generics was removed and the
generic Database interface was initialized with the Node type.

In addition, javadoc comments where added.

* Removed generics of OverflowDatabase in test classes

The generic type of OverflowDatabase was removed due to type dependencies. This
updates the test classes to not use generics.

* Use generic interface instead of implementing class

Changed type of field to the generic interface instead of the implementing
class. This hides accessible methods in OverflowDatabase.

* Save nodes from AST iteratively

Nodes from the AST are now iteratively traversed and submitted for creation.

* Trying to remove some code smells

* Remove recursion along graph edges

Previously, the database would persist a node by creating a corresponding vertex. Afterwards,
the edges of this node would be traversed recursively. Thus, all adjacent nodes would be
processed, their edges and their adjacent nodes. This aproach causes deep recursions that
cause stack overflows.

This new approach traverses only to adjacent nodes and returns them for subsequent iterative
processing. This reduces recursive depth.

Co-authored-by: Banse, Christian <christian.banse@aisec.fraunhofer.de>