Add builtin _split_match
_split_match is a new builtin. It takes a string, splits with a given regex and
checks if the resulting elements match the elements in the list specified as third
argument. The matching is set based, i.e. if both lists contain the same elements
independent of their position in the list.

This builtin helps to analyze arguments, where a whitelist must be enforced but the
order of elements is not fixed.