Pick up default location of generated sources (#212)
Used to change destination directory for generated sources by
annotation processors. This seemed to cause problems when e.g.
updating gradle. It required a manual delete of the generated
sources directory.

This keeps the default generated sources destination and adds
the directories to the source sets to be picked up by IDEs.