Resolve base of declared reference expressions when resolving MARK operands
This adds an additional step when resolving MARK operands. Operands might be
associated to declared reference expression. The MARK entity of the operand
might, however, be associated to a variable declaration. Previously, the
declared reference expression did not resolve to the variable declaration.
As a result, Codyze could not resolve a base. This change now resolves to the
referenced declaration.