Add builtin split and disjoint
It takes as the first argument a string, splits it using the string of
the second argument and checks, if the resulting list of elements is
disjoint from the list provided as third argument.

This builtin becomes useful if one wants to check that some options/
configurations are not set.