Replacing OverflowDB with CPG in-memory Graph
This is the start of the removal of OverflowDB. I will do this gradually and trying not to brake too many things at once. The first step is to move the Evaluator into a LegacyEvaluator and start with a new Evaluator class.

This is slightly blocked with support for CPG 4.0 (see #184), because otherwise we cannot add new features to the CPG graph if we need them.

Closes #218