Refactor builtin's execute method for wider return types
Previously, builtin's execute method was limited to ConstantValue. This
meant that builtins could only return booleans, numbers or strings. It
wasn't possible to return lists. This was an undue limitation on the
possibilities of builtins.

This refactors the return type of execute to the more general
MarkIntermediateResult. This should allow to return ListValue for further
processing.