Change evaluation of `when` to fail on errors and fix rule mistakes (#258)
* Change evaluation of `when` to fail on errors and fix rule mistakes

The `when` prevents the evaluation of the `ensure` in a rule. It acts as a
guard. Errors in the evaluation of the `when` clause used to be ignored. This
caused an evaluation of the `ensure` clause.

This changes the behaviour to fail and ignore the `ensure` clause. Thus,
errors in `when` are now considered to mean `false`. This prevents the
unnecessary evaluation of rules for example when a corresponding node
wasn't found.

As a side effect, mistakes in rules have been identified and fixed as well.

* Remove old code removing code smells