Reconfigure version range for vscode
Set engine and @types/vscode back to version range. Both are set
to the same range value.

This should improve usability of Codyze VSCode plugin as exact
version matches are removed. The exact version causes problems
when VSCode releases outpace the version demanded by the plugin.
This has prevented the installation of the plugin in newer
VSCode instances.