Change rule evaluation to create finding on errors (#356)
* Change rule evaluation to create finding on errors

When an evaluation of a rule returns an error, a finding is generated. The
finding is a rule violation. This change is more appropriate to reflect
the result of a rule evaluation. On an error, some part of a rule couldn't
be analyzed. Thus, it's unknown if a true violation exists. However,
ignoring the error may suggest that a rule is valid. This may be the more
severe conclusion. In addition, rules with errors need to be checked
manually to ensure compliance.

This change adjust unit test that changed due to the change in rule
evaluation.

* Quick fix possible heisenbug