Separate ServerConfiguration and TranslationConfiguration (#428)
This PR changes the AnalysisServer to use the buildTranslationConfiguration and buildServerConfiguration function of the Configuration class to build the TranslationConfiguration and ServerConfiguration object. All CPG configurations are removed from the ServerConfiguration class.

Codyze configurations are separated into the ones used by the AnalysisServer (stored in CodyzeConfiguration class) and the ones used in the Main class (stored in Configuration class). The server and translation configurations can only be accessed through calling the build...Configuration function of the corresponding configuration.